requirements.txt
anyio==4.11.0
certifi==2025.11.12
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.11
python-telegram-bot==20.0
rfc3986==1.5.0
sniffio==1.3.1
